4.44.0
* Added: Initial re-implementation of emote information cards. These have been broken for a while due to changes in Twitch's website preventing us from accessing them.
* Added: Setting to hide charity progress elements in chat.
* Changed: Emote tool-tips now display emote artists.
* Changed: The `Chat Command` chat action can now optionally be sent in another channel. Note that when doing so, you will not receive feedback from your sent message.
* Fixed: The 'Viewer Count' tool-tip duplicating itself.
* Fixed: Emote menu repeatedly requesting FFZ data.

* API Added: Ephemeral profiles can now be created by passing `ephemeral: true` in the options when creating a profile. These profiles are temporary and read-only.

+++ src/modules/main_menu/setting-mixin.js
@@ -96,6 +96,10 @@ export default {
 			return this.isDefault || this.validate(this.value);
 		},
 
+		isReadOnly() {
+			return this.profile && this.profile.ephemeral
+		},
+
 		sourceOrder() {
 			return this.source ? this.source.order : Infinity
 		},
@@ -210,6 +214,8 @@ export default {
 
 		set(value) {
 			// TODO: Run validation.
+			if ( this.isReadOnly )
+				return;
 
 			let process = this.item.process;
 			if ( process ) {
@@ -228,6 +234,9 @@ export default {
 		},
 
 		clear() {
+			if ( this.isReadOnly )
+				return;
+
 			this.profile.delete(this.item.setting);
 
 			if ( this.item.onUIChange )
